requirements.txt
fastapi==0.115.0
boto3==1.35.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
aiosqlite==0.20.0
asyncpg==0.30.0
httpx==0.27.0
anthropic==0.39.0
feedparser==6.0.11
python-dotenv==1.0.1
pydantic-settings==2.5.0
python-multipart==0.0.22
jinja2==3.1.4
beautifulsoup4==4.12.3
lxml==5.3.0
google-auth
google-auth-oauthlib
google-api-python-client
openai
google-generativeai
PyJWT>=2.8.0
cryptography>=41.0.0
supabase>=2.0.0
